svn:revision:71099
svn:branch:brep-debug
svn:account:n_reed

+++ include/bg/trimesh.h
@@ -68,9 +68,9 @@ BG_EXPORT extern void bg_free_trimesh_edges(struct bg_trimesh_edges *edges);
 BG_EXPORT extern void bg_free_trimesh_faces(struct bg_trimesh_faces *faces);
 BG_EXPORT extern void bg_free_trimesh_solid_errors(struct bg_trimesh_solid_errors *errors);
 
-BG_EXPORT extern int bg_trimesh_closed_fan(size_t vcnt, size_t fcnt, fastf_t *v, int *f);
-BG_EXPORT extern int bg_trimesh_orientable(size_t vcnt, size_t fcnt, fastf_t *v, int *f);
-BG_EXPORT extern int bg_trimesh_solid(size_t vcnt, size_t fcnt, fastf_t *v, int *f, int **bedges);
+BG_EXPORT extern int bg_trimesh_closed_fan(int vcnt, int fcnt, fastf_t *v, int *f);
+BG_EXPORT extern int bg_trimesh_orientable(int vcnt, int fcnt, fastf_t *v, int *f);
+BG_EXPORT extern int bg_trimesh_solid(int vcnt, int fcnt, fastf_t *v, int *f, int **bedges);
 
 /* The below functions are for use as arguments to error tests. Given
  * a face/edge, they return true if the caller should continue
@@ -82,12 +82,12 @@ BG_EXPORT extern int bg_trimesh_solid(size_t vcnt, size_t fcnt, fastf_t *v, int 
  * pre-allocated members of the correct size and count members set to
  * 0, that they will populate.
  */
-typedef int (*bg_face_error_func_t)(size_t face_idx, void *data);
+typedef int (*bg_face_error_func_t)(int face_idx, void *data);
 typedef int (*bg_edge_error_funct_t)(struct bg_trimesh_halfedge *edge, void *data);
 
-BG_EXPORT extern int bg_trimesh_face_exit(size_t face_idx, void *data);
-BG_EXPORT extern int bg_trimesh_face_continue(size_t face_idx, void *data);
-BG_EXPORT extern int bg_trimesh_face_gather(size_t face_idx, void *data);
+BG_EXPORT extern int bg_trimesh_face_exit(int face_idx, void *data);
+BG_EXPORT extern int bg_trimesh_face_continue(int face_idx, void *data);
+BG_EXPORT extern int bg_trimesh_face_gather(int face_idx, void *data);
 BG_EXPORT extern int bg_trimesh_edge_exit(struct bg_trimesh_halfedge *edge, void *data);
 BG_EXPORT extern int bg_trimesh_edge_continue(struct bg_trimesh_halfedge *edge, void *data);
 BG_EXPORT extern int bg_trimesh_edge_gather(struct bg_trimesh_halfedge *edge, void *data);
@@ -96,13 +96,13 @@ BG_EXPORT extern int bg_trimesh_edge_gather(struct bg_trimesh_halfedge *edge, vo
  * Otherwise, they return the number of instances of the error found
  * before the error function argument returned false (at least 1).
  */
-BG_EXPORT extern int bg_trimesh_degenerate_faces(size_t num_faces, int *fpoints, bg_face_error_func_t degenerate_func, void *data);
-BG_EXPORT extern int bg_trimesh_unmatched_edges(size_t num_edges, struct bg_trimesh_halfedge *edge_list, bg_edge_error_funct_t error_edge_func, void *data);
-BG_EXPORT extern int bg_trimesh_misoriented_edges(size_t num_edges, struct bg_trimesh_halfedge *edge_list, bg_edge_error_funct_t error_edge_func, void *data);
-BG_EXPORT extern int bg_trimesh_excess_edges(size_t num_edges, struct bg_trimesh_halfedge *edge_list, bg_edge_error_funct_t error_edge_func, void *data);
-BG_EXPORT extern int bg_trimesh_solid2(size_t vcnt, size_t fcnt, fastf_t *v, int *f, struct bg_trimesh_solid_errors *errors);
+BG_EXPORT extern int bg_trimesh_degenerate_faces(int num_faces, int *fpoints, bg_face_error_func_t degenerate_func, void *data);
+BG_EXPORT extern int bg_trimesh_unmatched_edges(int num_edges, struct bg_trimesh_halfedge *edge_list, bg_edge_error_funct_t error_edge_func, void *data);
+BG_EXPORT extern int bg_trimesh_misoriented_edges(int num_edges, struct bg_trimesh_halfedge *edge_list, bg_edge_error_funct_t error_edge_func, void *data);
+BG_EXPORT extern int bg_trimesh_excess_edges(int num_edges, struct bg_trimesh_halfedge *edge_list, bg_edge_error_funct_t error_edge_func, void *data);
+BG_EXPORT extern int bg_trimesh_solid2(int vcnt, int fcnt, fastf_t *v, int *f, struct bg_trimesh_solid_errors *errors);
 
-BG_EXPORT extern struct bg_trimesh_halfedge * bg_trimesh_generate_edge_list(size_t fcnt, int *f);
+BG_EXPORT extern struct bg_trimesh_halfedge * bg_trimesh_generate_edge_list(int fcnt, int *f);
 
 __END_DECLS
